more sound

+++ src/engine/sound/sound.cpp
@@ -1,7 +1,7 @@
 #include "sound.hpp"
 
 Sound::Sound(WAVEFormat fmt, int size, char* data)
-	: fmt{fmt}, length{(float)size / (fmt.numChannels * fmt.sampleRate * (fmt.bitsPerSample / 8.0f)) * 1000.0f}, size{size}, data{data} {
+	: fmt{fmt}, length{size / (fmt.numChannels * fmt.sampleRate * (fmt.bitsPerSample / 8.0f)) * 1000.0f}, size{size}, data{data} {
 	alGenBuffers(1, &buffer);
 	alBufferData(buffer, OALFormat(), data, size, (ALsizei)fmt.sampleRate);
 }
@@ -9,7 +9,6 @@ Sound::Sound(WAVEFormat fmt, int size, char* data)
 ALenum Sound::OALFormat() {
 	if (fmt.bitsPerSample == 16) {
 		return fmt.numChannels == 2 ? AL_FORMAT_STEREO16 : AL_FORMAT_MONO16;
-
 	} else if (fmt.bitsPerSample == 8) {
 		return fmt.numChannels == 2 ? AL_FORMAT_STEREO8 : AL_FORMAT_MONO8;
 	}
